Fix bugs and warning when building on Linux

+++ ncfw/include/ncfw/Algebra/Angle.hpp
@@ -12,9 +12,9 @@ public:
 
   float rad() const { return m_rad; }
   float deg() const { return 180.0f * m_rad / (float)M_PI; }
-  float sin() const { return std::sinf(m_rad); }
-  float cos() const { return std::cosf(m_rad); }
-  float tan() const { return std::tanf(m_rad); }
+  float sin() const { return std::sin(m_rad); }
+  float cos() const { return std::cos(m_rad); }
+  float tan() const { return std::tan(m_rad); }
 
   Angle operator+(const Angle &p_rhs) const { return {m_rad + p_rhs.m_rad}; }
   Angle operator-(const Angle &p_rhs) const { return {m_rad - p_rhs.m_rad}; }
@@ -28,7 +28,7 @@ public:
 private:
   float m_rad;
 
-  Angle(float p_rad) : m_rad(p_rad - 2.0f * M_PI * std::floorf(p_rad / (2.0f * M_PI))) {}
+  Angle(float p_rad) : m_rad(p_rad - 2.0f * M_PI * std::floor(p_rad / (2.0f * M_PI))) {}
 };
 
 inline Angle operator*(float p_lhs, const Angle &p_rhs) { return Angle::rad(p_lhs * p_rhs.rad()); }
